[REF] web_tour: harmonize findTrigger between tour_interactive and tour_automatic
Move the trigger resolution rules (modal/overlay stacking, UI-blocked,
enabled, parent frame readiness) from TourStepAutomatic up to the shared
TourStep base, so manual tours honor the same rules as automatic ones.
